fix(naming): renames ts file in camelCase and refactors imports
